Numerical Kinematics introduced via Momentum Twistors
The package SpinorFunctions.wl has become SpinorHelicity.wl, which is the main .wl file.

The tilded and untilded masses in four dimensions have been introduced for numerical purposes.

The NumericalKinematics.wl contain a series of function to generate rational kinematics via Momentum Twistors, for massive particles (even identical masses) as well.

The function NEvaluate has been added to SpinorHelicity.wl, because it was not working within NumericalKinematics. In that case, Block defined the variables in the context NumericalKinematics`Private`, which is unpleasant.